[workspace-tools] Move git helpers to separate files
