+++ packages/workspace-tools/src/__tests__/git/getRecentCommitMessages.test.ts
@@ -3,7 +3,7 @@ import { cleanupFixtures, setupFixture } from "../setupFixture.js";
 import fs from "fs";
 import path from "path";
 import { git as _git, type GitOptions } from "../../git/git.js";
-import { getRecentCommitMessages } from "../../git/gitUtilities.js";
+import { getRecentCommitMessages } from "../../git/getRecentCommitMessages.js";
 
 /** Call git helper but throw on error by default */
 const git = (args: string[], opts: GitOptions) => _git(args, { throwOnError: true, ...opts });
