+++ packages/workspace-tools/src/__tests__/git/getUnstagedChanges.test.ts
@@ -3,7 +3,7 @@ import { cleanupFixtures, setupFixture } from "../setupFixture.js";
 import fs from "fs";
 import path from "path";
 import { git as _git, type GitOptions } from "../../git/git.js";
-import { getUnstagedChanges } from "../../git/gitUtilities.js";
+import { getUnstagedChanges } from "../../git/getChanges.js";
 
 /** Call git helper but throw on error by default */
 const git = (args: string[], opts: GitOptions) => _git(args, { throwOnError: true, ...opts });
