+++ packages/workspace-tools/src/__tests__/git/getUntrackedChanges.test.ts
@@ -3,7 +3,7 @@ import { cleanupFixtures, setupFixture } from "../setupFixture.js";
 import fs from "fs";
 import path from "path";
 import { git as _git, type GitOptions } from "../../git/git.js";
-import { getUntrackedChanges } from "../../git/gitUtilities.js";
+import { getUntrackedChanges } from "../../git/getChanges.js";
 import type { GitCommonOptions } from "../../git/types.js";
 
 /** Call git helper but throw on error by default */
